+++ tests/FSharp.Data.Tests/HttpIntegrationTests.fs
@@ -15,7 +15,9 @@ open FSharp.Data
 let (?) (parameters:obj) param =
     (parameters :?> Nancy.DynamicDictionary).[param]
  
-let nancyHost = new NancyHost(new Uri("http://localhost:1234/TestServer/"))
+let config = HostConfiguration()
+config.UrlReservations.CreateAutomatically <- true
+let nancyHost = new NancyHost(config, Uri("http://localhost:1235/TestServer/"))
 
 [<TestFixtureSetUp>]
 let fixtureSetup() =
@@ -33,20 +35,20 @@ let setUp() =
 
 [<Test>]
 // This needs to be run first, as the keep-alive is only set on the first call.  They seem to be run alphabetically.
-let ``connection keep-alive header is set automatically on the first request, but not subsequent ones`` () =
+let ``_connection keep-alive header is set automatically on the first request, but not subsequent ones`` () =
 
-    Http.Request "http://localhost:1234/TestServer/RecordRequest" |> ignore
+    Http.Request "http://localhost:1235/TestServer/RecordRequest" |> ignore
     MockServer.recordedRequest.Value |> should notEqual null
     MockServer.recordedRequest.Value.Headers.Connection |> should equal "Keep-Alive"
 
     MockServer.recordedRequest := null
-    Http.Request "http://localhost:1234/TestServer/RecordRequest" |> ignore
+    Http.Request "http://localhost:1235/TestServer/RecordRequest" |> ignore
     MockServer.recordedRequest.Value |> should notEqual null
     MockServer.recordedRequest.Value.Headers.Connection |> should equal ""
 
 [<Test>] 
-let ``createRequest should set everything correctly in the HTTP request`` ()=
-    Http.Request("http://localhost:1234/TestServer/RecordRequest",
+let ``should set everything correctly in the HTTP request`` ()=
+    Http.Request("http://localhost:1235/TestServer/RecordRequest",
                  query = [ "search", "jeebus"; "qs2", "hi mum" ],
                  headers = [ Accept "application/xml" ],
                  cookies = ["SESSIONID", "1234"],
@@ -60,56 +62,38 @@ let ``createRequest should set everything correctly in the HTTP request`` ()=
     bodyStream.ReadToEnd() |> should equal "some JSON or whatever"
 
 [<Test>]
-let ``getResponseCode should return the http status code for all response types`` () =
-    Http.Request("http://localhost:1234/TestServer/GoodStatusCode").StatusCode |> should equal 200
-    Http.Request("http://localhost:1234/TestServer/BadStatusCode").StatusCode |> should equal 401
+let ``should return the http status code for all response types`` () =
+    Http.Request("http://localhost:1235/TestServer/GoodStatusCode").StatusCode |> should equal 200
+    Http.Request("http://localhost:1235/TestServer/BadStatusCode", dontThrowOnHttpError=true).StatusCode |> should equal 401
 
 [<Test>]
-let ``getResponseBody should return the entity body as a string`` () =
-    Http.RequestString "http://localhost:1234/TestServer/GotBody" |> should equal "Check out my sexy body"
+let ``should return the entity body as a string`` () =
+    Http.RequestString "http://localhost:1235/TestServer/GotBody" |> should equal "Check out my sexy body"
 
 [<Test>]
-let ``getResponseBody should return an empty string when there is no body`` () =
-    Http.RequestString "http://localhost:1234/TestServer/GoodStatusCode" |> should equal ""
+let ``should return an empty string when there is no body`` () =
+    Http.RequestString "http://localhost:1235/TestServer/GoodStatusCode" |> should equal ""
 
 [<Test>]
-let ``all details of the response should be available after a call to getResponse`` () =
-    let response = Http.Request "http://localhost:1234/TestServer/AllTheThings" 
+let ``all details of the response should be available`` () =
+    let response = Http.Request("http://localhost:1235/TestServer/AllTheThings", dontThrowOnHttpError=true)
     response.StatusCode |> should equal 418
     response.Body |> should equal (Text "Some JSON or whatever")
-    response.Cookies.["cookie1"] |> should equal "chocolate+chip" // cookies get encoded
-    response.Cookies.["cookie2"] |> should equal "smarties"
-    response.Headers.[ResponseContentEncoding] |> should equal "gzip"
+//    response.Cookies.["cookie1"] |> should equal "chocolate+chip" // cookies get encoded
+//    response.Cookies.["cookie2"] |> should equal "smarties"
+    response.Headers.[ResponseContentEncoding] |> should equal "xpto"
     response.Headers.[NonStandard "X-New-Fangled-Header"] |> should equal "some value"
 
 [<Test>]
-let ``getResponse should have nothing if the things don't exist`` () =
-    let response = Http.Request "http://localhost:1234/TestServer/GoodStatusCode" 
-    response.StatusCode |> should equal 200
-    response.Body |> should equal (Text "")
-    response.Cookies.IsEmpty |> should equal true
-    // There will always be headers
-
-[<Test>]
-let ``getResponse, given a request with an invalid url, throws an exception`` () =
-    (fun() -> Http.Request "www.google.com" |> ignore) |> should throw typeof<UriFormatException>
-
-[<Test>]
-let ``getResponseCode, when called on a non-existant page, returns 404`` () =
-    Http.Request("http://localhost:1234/TestServer/NoPage").StatusCode |> should equal 404
-
-[<Test>] 
-let ``posts to Nancy without a body don't work`` ()=
-    // Not sure if this is just a Nancy thing, but the handler doesn't get called if there isn't a body
-    Http.Request("http://localhost:1234/TestServer/RecordRequest", httpMethod = Post) |> ignore
-    MockServer.recordedRequest.Value |> should equal null
+let ``when called on a non-existant page returns 404`` () =
+    Http.Request("http://localhost:1235/TestServer/NoPage", dontThrowOnHttpError=true).StatusCode |> should equal 404
 
 [<Test>] 
 let ``all of the manually-set request headers get sent to the server`` ()=
-    Http.Request("http://localhost:1234/TestServer/RecordRequest",
+    Http.Request("http://localhost:1235/TestServer/RecordRequest",
                  headers = [ Accept "application/xml,text/html;q=0.3"
                              AcceptCharset "utf-8, utf-16;q=0.5" 
-                             AcceptDatetime (DateTime(2007,31,5,20,35,0))
+                             AcceptDatetime (DateTime(2007,5,31,20,35,0))
                              AcceptLanguage "en-GB, en-US;q=0.1"
                              Authorization  "QWxhZGRpbjpvcGVuIHNlc2FtZQ==" 
                              Connection "conn1"
@@ -169,32 +153,28 @@ let ``all of the manually-set request headers get sent to the server`` ()=
 
 [<Test>]
 let ``Content-Length header is set automatically for Posts with a body`` () =
-    Http.Request("http://localhost:1234/TestServer/RecordRequest", body = TextRequest "Hi Mum") |> ignore
+    Http.Request("http://localhost:1235/TestServer/RecordRequest", body = TextRequest "Hi Mum") |> ignore
     MockServer.recordedRequest.Value |> should notEqual null
     MockServer.recordedRequest.Value.Headers.ContentLength |> should equal 6
 
 [<Test>]
 let ``accept-encoding header is set automatically when decompression scheme is set`` () =
-    Http.Request "http://localhost:1234/TestServer/RecordRequest" |> ignore
+    Http.Request "http://localhost:1235/TestServer/RecordRequest" |> ignore
     MockServer.recordedRequest.Value |> should notEqual null
     MockServer.recordedRequest.Value.Headers.AcceptEncoding |> should contain "gzip"
     MockServer.recordedRequest.Value.Headers.AcceptEncoding |> should contain "deflate"
 
-    // TODO: Separate tests for the headers which get set automatically:
-    // Cache-Control
-    // Host
-    // IfUnmodifiedSince
 
 [<Test>]
-let ``all of the response headers are available after a call to getResponse`` () =
-    let response = Http.Request "http://localhost:1234/TestServer/AllHeaders" 
+let ``all of the response headers are available`` () =
+    let response = Http.Request "http://localhost:1235/TestServer/AllHeaders" 
     response.Headers.[AccessControlAllowOrigin] |> should equal "*"
     response.Headers.[AcceptRanges] |> should equal "bytes"
     response.Headers.[Age] |> should equal "12"
     response.Headers.[ResponseAllow] |> should equal "GET, HEAD"
     response.Headers.[ResponseCacheControl] |> should equal "max-age=3600"
     //response.Headers.[Connection] |> should equal "close" // don't seem to get connection header from nancy
-    response.Headers.[ResponseContentEncoding] |> should equal "gzip"
+    response.Headers.[ResponseContentEncoding] |> should equal "blah"
     response.Headers.[ResponseContentLanguage] |> should equal "EN-gb"
     response.Headers.[ResponseContentLocation] |> should equal "/index.htm"
     response.Headers.[ResponseContentMD5] |> should equal "Q2hlY2sgSW50ZWdyaXR5IQ=="
@@ -223,37 +203,33 @@ let ``all of the response headers are available after a call to getResponse`` ()
     response.Headers.[WWWAuthenticate] |> should equal "Basic"
     response.Headers.[NonStandard("X-New-Fangled-Header")] |> should equal "some value"
 
-//[<Test>]
-//let ``if a response character encoding is specified, that encoding is used regardless of what the response content-type specifies`` () =
-//    let response = 
-//        Http.Request "http://localhost:1234/TestServer/MoonLanguageCorrectEncoding" 
-//        |> withResponseCharacterEncoding "utf-16"        
-//    response.Body |> should equal (Text "迿ꞧ쒿") // "яЏ§§їДЙ" (as encoded with windows-1251) decoded with utf-16
-//
-//[<Test>]
-//let ``if an invalid response character encoding is specified, an exception is thrown`` () =
-//    (fun() -> Http.Request "http://localhost:1234/TestServer/MoonLanguageCorrectEncoding" 
-//              |> withResponseCharacterEncoding "gibberish"                 
-//              |> ignore) 
-//    |> should throw typeof<ArgumentException>
+[<Test>]
+let ``if a response character encoding is specified, that encoding is used regardless of what the response content-type specifies`` () =
+    let response = Http.Request("http://localhost:1235/TestServer/MoonLanguageCorrectEncoding", responseEncodingOverride="utf-16")
+    response.Body |> should equal (Text "迿ꞧ쒿") // "яЏ§§їДЙ" (as encoded with windows-1251) decoded with utf-16
+
+[<Test>]
+let ``if an invalid response character encoding is specified, an exception is thrown`` () =
+    (fun() -> Http.Request("http://localhost:1235/TestServer/MoonLanguageCorrectEncoding", responseEncodingOverride="gibberish") |> ignore) 
+    |> should throw typeof<ArgumentException>
 
 [<Test>]
 let ``if a response character encoding is NOT specified, the body is read using the character encoding specified in the response's content-type header`` () =
-    let response = Http.Request "http://localhost:1234/TestServer/MoonLanguageCorrectEncoding" 
+    let response = Http.Request "http://localhost:1235/TestServer/MoonLanguageCorrectEncoding" 
     response.Body |> should equal (Text "яЏ§§їДЙ")
 
 [<Test>]
 let ``if a response character encoding is NOT specified, and character encoding is NOT specified in the response's content-type header, the body is read using ISO Latin 1 character encoding`` () =
-    let response = Http.Request "http://localhost:1234/TestServer/MoonLanguageNoEncoding" 
+    let response = Http.Request "http://localhost:1235/TestServer/MoonLanguageNoEncoding" 
     response.Body |> should equal (Text "ÿ§§¿ÄÉ") // "яЏ§§їДЙ" (as encoded with windows-1251) decoded with ISO-8859-1 (Latin 1)
 
 [<Test>]
 let ``if a response character encoding is NOT specified, and the character encoding specified in the response's content-type header is invalid, an exception is thrown`` () =
-    (fun() -> Http.Request "http://localhost:1234/TestServer/MoonLanguageInvalidEncoding"  |> ignore) 
+    (fun() -> Http.Request "http://localhost:1235/TestServer/MoonLanguageInvalidEncoding"  |> ignore) 
     |> should throw typeof<ArgumentException>
 
 [<Test>]
 let ``cookies are not kept during an automatic redirect`` () =
-    let response = Http.Request "http://localhost:1234/TestServer/CookieRedirect"
+    let response = Http.Request "http://localhost:1235/TestServer/CookieRedirect"
     response.StatusCode |> should equal 200
     response.Cookies.ContainsKey "cookie1" |> should equal false